CMR-72: Add deletes task to ingest dag.
This is just an idea that depends on implementing
a /records/api/deletes endpoint that accepts a marcfile and knows how
to delete the records in that file as proposed in REF:
https://tulibdev.atlassian.net/browse/CMR-73.

Alternatively another option is to process the marc file
direclty in the dag but that would spread the complexity of managing
the marcfile into the dag as well as the API which I think is harder
to reason through than posting a file to the API and have it do what it
knows what to do with it.

Another idea is to have the API service just send us a list of record ids given a marc file with multiple records to be deleted.
And then we iterate over the retrieved list and send DELETE
requests per record id to records/:id.

But this makes the ingest and delete dag tasks more different than each
other whereas if we just POST the file to be deleted we can keep the
ingest script and the deletes script basically identical other than what
files get sent and what endpoint gets hit.